feat: move shuttle button to components folder
