--- src/pages/Shuttle/ConfirmModal/ShuttleButton/index.js
@@ -1,2 +1,0 @@
-export {default as ShuttleInButton} from './In'
-export {default as ShuttleOutButton} from './Out'
