+++ src/pages/components/index.js
@@ -15,3 +15,5 @@ export {default as MobileFooter} from './MobileFooter'
 export {default as useTransactionNotification} from './useTransactionNotification'
 export {default as RiskModal} from './RiskModal'
 export {default as AccountError} from './AccountError'
+export {default as ShuttleInButton} from './ShuttleButton/In'
+export {default as ShuttleOutButton} from './ShuttleButton/Out'
